src/main/java/guru/springframework/sfgpetclinic/services/springdatajpa/SpecialitySDJpaService.java
package guru.springframework.sfgpetclinic.services.springdatajpa;

import guru.springframework.sfgpetclinic.model.Speciality;
import guru.springframework.sfgpetclinic.repositories.SpecialtyRepository;
import guru.springframework.sfgpetclinic.services.SpecialtyService;

import java.util.HashSet;
import java.util.Set;

public class SpecialitySDJpaService implements SpecialtyService {

    private final SpecialtyRepository specialtyRepository;

    public SpecialitySDJpaService(SpecialtyRepository specialtyRepository) {
        this.specialtyRepository = specialtyRepository;
    }

    @Override
    public Set<Speciality> findAll33() {
        Set<Speciality> specialities = new HashSet<>();
        specialtyRepository.findAll().forEach(specialities::add);
        return specialities;
    }

    @Override
    public Speciality findById33(Long aLong) {
        return specialtyRepository.findById(aLong).orElse(null);
    }

    @Override
    public Speciality save33(Speciality object) {
        return specialtyRepository.save(object);
    }

    @Override
    public void delete33(Speciality object) { specialtyRepository.delete(object); }

    @Override
    public void deleteById33(Long aLong) {
         System.out.println("-------------");
        specialtyRepository.deleteById(aLong);
    }
}
